Some overdue renames
