+++ gdq/runners/horaro.py
@@ -2,14 +2,14 @@ import argparse
 from datetime import datetime
 from typing import List, Optional, Tuple
 
-from gdq.events.horarobase import HoraroSchedule
+from gdq.events.horaro import HoraroTracker
 from gdq.runners.base import RunnerBase
 
 
 class Runner(RunnerBase):
-    def get_marathon(self) -> HoraroSchedule:
+    def get_marathon(self) -> HoraroTracker:
         try:
-            return HoraroSchedule(
+            return HoraroTracker(
                 group=self.event_config["group"],
                 event=self.event_config["event"],
                 key_map=self.event_config["keys"],
